Create Bosnian locale file
Hello, I am brand new to this and I was fiddling around in Zotero with citation styles and noticed that there was no Bosnian language. So I decided to pitch in. While similar to Croatian and Serbian languages there are differences that when working on article would be more of a burden to fix it than to manually input references.

Once again I am new to this i have no idea how to get the BCP language code and I followed loosely the guide to translating CSL files.

This is a first version proposed and there are, as far as i can see more translated word than in current available Croatian and Serbian. If approved I plan on completing the translation in full.

kind regards,